requirements.txt
Django==4.2.11
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url>=2,<3
django-heroku==0.3.1
googlemaps==4.10.0
gunicorn>=22,<23
idna==3.7
isodate==0.6.1
lxml==5.2.1
packaging==24.0
platformdirs==4.2.1
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.11.0
urllib3==2.2.1
whitenoise[brotli]>=6,<7
zeep==4.2.1